Make JSON validator a separate component
